requirements.txt
tensorflow-gpu>=2.0,<=2.2
graphviz
numpy
scipy
sympy
iminuit
opt_einsum
matplotlib
PyYAML
uproot
